==> 0.train_val_split.ipynb <==
## Code to split training set into two sets : training and validation clinical notes datasets
NB one patient with two or more different clinical notes should not appear in both groups

from os.path import isfile, isdir, join, basename
from os import listdir
import pandas as pd
import numpy as np
import re
# import matplotlib.pyplot as plt
import collections
import math
# from sklearn import metrics
import random
import shutil 

IN_TXT_DIR = '/export/home/cse200093/Expe_Pheno/data/expe_pheno_back_up_26_11'

assert isdir(IN_TXT_DIR)


# Read all text files in the dir and split two random datasets : train and val

# extract all ann_files from IN_BRAT_DIR
txt_files = [f for f in listdir(IN_TXT_DIR) if isfile(join(IN_TXT_DIR, f)) if f.endswith('.txt')]

Ident = {}

for txt_file in txt_files:
    txt_path = join(IN_TXT_DIR, txt_file)
    
    # Read ann file
    with open(txt_path, 'r', encoding='utf-8') as f_in:
        lines = f_in.readlines()
        #print("\n\n text_path", txt_file[:-4])
        #print(lines[0])
        # keep track of the patient's NIP
        Ident[txt_file[:-4]]=lines[0]
        

# convert the dictionnary to a Pandas dataframe        
txt_df = pd.DataFrame(list(Ident.items()), columns = ['File', 'Id_Patient'])

# Count number of total clinical notes of the file
print("number of CRH", len(txt_df))

# count number of unique patients
print("Unique Patient", len(txt_df["Id_Patient"].unique()))


txt_df.sample(5)

# Get a set of all Id of patients : 
Id_unique = set(txt_df['Id_Patient'])
# Check the size
print("Set of unique Patients :", len(Id_unique))

# and get a random set of train and val : 
print('size of validation dataset =',int(17*len(txt_df)/100))
val_patients = random.sample(Id_unique, int(17*len(txt_df)/100))

with open('set_id_patient_expe_pheno.txt','w') as f:
    f.write(str(Id_unique))

import ast
with open('set_id_patient_expe_pheno.txt','r') as f:
    my_set = ast.literal_eval(f.read())

Id_unique == my_set

# check if it leds to the right number of CRH : 

# get the corresponding files names
list_Val = []
for key, value in Ident.items():
    if value in val_patients:
        # print('validation patient', value)
        list_Val.append(key)


list_Val_2 = [key for key,value in Ident.items() if value in val_patients]

len(list_Val), len(list_Val_2)

list_Val

# copy the corresponding validation files in the val directory
import shutil 
for i in list_Val: 
    # copy txt files
    source_file = f'/export/home/cse200093/Expe_Pheno/data/expe_pheno_back_up_26_11/{i}.txt'
    des_file = f'/export/home/cse200093/Expe_Pheno/data/final_val/{i}.txt'
    shutil.copyfile(source_file,des_file)
    # and they respective .ann files 
    source_ann_file = f'/export/home/cse200093/Expe_Pheno/data/expe_pheno_back_up_26_11/{i}.ann'
    des_ann_file = f'/export/home/cse200093/Expe_Pheno/data/final_val/{i}.ann'
    shutil.copyfile(source_ann_file,des_ann_file)
    
source_file

# Training files : 
list_Train = [key for key,value in Ident.items() if value not in val_patients]
len(list_Train)

# copy the corresponding training files in the train directory

for i in list_Train : 
    # copy txt files
    source_file = f'/export/home/cse200093/Expe_Pheno/data/expe_pheno_back_up_26_11/{i}.txt'
    des_file = f'/export/home/cse200093/Expe_Pheno/data/final_train/{i}.txt'
    shutil.copyfile(source_file,des_file)
    # and they respective .ann files 
    source_ann_file = f'/export/home/cse200093/Expe_Pheno/data/expe_pheno_back_up_26_11/{i}.ann'
    des_ann_file = f'/export/home/cse200093/Expe_Pheno/data/final_train/{i}.ann'
    shutil.copyfile(source_ann_file,des_ann_file)
    
source_file

txt_df.head()

# compute thenumber of patients for each disease
taka_df = txt_df[txt_df['File'].str.contains('taka')]
# count number of unique patients
print("Unique Takayasu Patient", len(taka_df["Id_Patient"].unique()))


lupus_df = txt_df[txt_df['File'].str.contains('lupus')]
# count number of unique patients
print("Unique Lupus Patient", len(lupus_df["Id_Patient"].unique()))


scl_df = txt_df[txt_df['File'].str.contains('scl')]
# count number of unique patients
print("Unique scl Patient", len(scl_df["Id_Patient"].unique()))

sapl_df = txt_df[txt_df['File'].str.contains('sapl')]
# count number of unique patients
print("Unique SAPL Patient", len(sapl_df["Id_Patient"].unique()))

taka_df
