==> 3.Post_processing_get_FastText_emb.ipynb <==
import numpy as np 
import pandas as pd 
import re
import fasttext
import pickle

OUT_PATH = './out/15_02_22_CRH_VAL_FastText.pkl'

# Export mutlilabel classification prediction 

Takes a pandas dataframe with sigmoid values and export as a pd dataframe with all labels for all terms in the files, and get the genders

In a second step, get all the couples term/label per file in a pandas Dataframe, keeping track of the proportion 

In a third step : get FasText embeddings for all terms keep them in a pandas Dataframe. 

# Get the results of the classification step: 

model_name = '20220209T092729_expe_pheno_classifier_camembert_l_umls_fr_local'

result_path = f'./out/final_CRH_VAL_PHENO_pred_{model_name}.csv'
#result_path = 'final_test_SUB_EXAM_20210701T120111_classifier_camembert_umls_fr_mesh_mdr_val_0.1_deft.csv'
thresholds_path = f'./out/thresholds_{model_name}.csv'
labels_path = f'./out/labels_{model_name}.csv'

thresholds = np.loadtxt(thresholds_path, delimiter=",")
print(f'Thresholds loaded from {thresholds_path}')
with open(labels_path, 'r') as f_out:
    label_names = f_out.readline().strip().split(',')
print(f'Label names loaded from {labels_path}')
    
result_df = pd.read_csv(result_path)
if "Unnamed: 0" in result_df.columns:
    del(result_df["Unnamed: 0"])

print(f'Results loaded from {result_path}')

labels_from_results = [l for l in result_df.columns if l.startswith('label_')]
assert labels_from_results == label_names

result_data = result_df[label_names].to_numpy()
assert len(result_data[0]) == len(label_names) == len(thresholds)
# if all thresholds are the same, keep only one value
if len(set(thresholds)) == 1:
    thresholds = thresholds[0]



result_df = pd.read_csv(result_path)

# Convert them in a dataframe

seuils = [0.75]
#seuils = [0.90]
for seuil in seuils : 
    print("\n seuil =", seuil)
    # Apply thresholds on data
    result_data_th = result_data > seuil
    result_df[label_names] = result_data_th    
    df2 = result_df.copy()

df2.sample(5)

new_df = df2[df2==1].stack().reset_index().drop(0,1)
new_df

new_df ['term'] = range(len(new_df))
new_df ['source'] = range(len(new_df))

for i in range(len(new_df)):
    idx_init = new_df['level_0'][i]
    new_df['term'][i] = df2['term'][idx_init]
    new_df['source'][i] = df2['source'][idx_init]

new_df.sample(5)

new_df_corr = new_df.replace('label_', '', regex = True)
new_df_corr

# Get f and m terms to get the gender info 
df_gender = new_df_corr[(new_df_corr['term'] == 'f') | (new_df_corr['term'] == 'm')]
df_gender.sample(5)

del(df_gender['level_0'])
del(df_gender['level_1'])
df_gender.columns = ['gender', 'source']
df_gender.sample(5)

len(df_gender), len(set(df_gender.source.tolist()))

# get files without gender detection :
no_g = set(new_df_corr.source.tolist()) - set(df_gender.source.tolist())

# some patients have more than one gender detected : choose the gender that appears more times 
import collections

df_f = df_gender[df_gender['gender'] =='f']
df_m = df_gender[df_gender['gender'] =='m']
# get set of file names with two genders detected :
both_g = set(df_f.source.tolist()) & set(df_m.source.tolist())
counter_g = dict()
true_g = dict()
for file in both_g:
    # get the list of genders associated to this file 
    genders = df_gender[df_gender['source']== file].gender.tolist()
    counter_g[file] = collections.Counter(genders)
    if counter_g[file]['m'] >= counter_g[file]['f']:
        true_g[file] = 'm'
    else: 
        true_g[file] = 'f'
    # drop this from df_gender 
    df_gender = df_gender.drop(df_gender[df_gender['source'] == file].index)

# counter_g, true_g

# build the final gender dataset to be merged later on:
data_g = df_gender.values.tolist()
for file in both_g:
    data_g.append([true_g[file], file])
    
# given our population, by default, patient with no gender detected will be assigned to the female gender: 
for file_b in no_g:
    data_g.append(['f', file_b])

# convert the data into a dataframe 
columns = ['gender', 'source']
df_gender_corr = pd.DataFrame(data=list(data_g), columns=columns)
df_gender_corr = df_gender_corr.drop_duplicates().reset_index(drop=True)
len(df_gender_corr)

new_df_corr.drop(new_df_corr[(new_df_corr.term == 'f') | (new_df_corr.term == 'm')].index, inplace = True)

df_final = pd.merge(df_gender_corr, new_df_corr, on='source')
df_check = df_final.groupby('source')['level_1'].nunique()
len(df_check)

df_final
del(df_final['level_0'])

df_final.columns = ['gender', 'source','label', 'term']
df_final = df_final.replace('NEW', 'n', regex = True)
df_final.sample(5)

# Last step : Get the FastText embeddings of extracted and classified terms

model = fasttext.load_model("/export/home/cse200093/output/output.bin")

df_final.values[0,3]

data = []
for i in range(len(df_final.values)):
    data.append([df_final.values[i,1],df_final.values[i,0], df_final.values[i,3],
                 model.get_word_vector(df_final.values[i,3]), 
                df_final.values[i,2]])

# save this dataset in a pd dataFrame 
df = pd.DataFrame(data)

df.columns = ['source', 'gender','term', 'fastext_embeddings', 'label']
df.sample(5)

# save 
df.to_pickle(OUT_PATH)

# Finaly, in order to keep track of the importance ("weight") of a label at the document scale
# get the ratio : #term_for_a_given_label/#total_term

# empty list of list to "append" each values
data_C = []
# get the set of all documents source
enum = set(my_df.source)
for i in enum:
    # get the index of all terms for one  given document: 
    idx_i = my_df[my_df['source'] == i].index.tolist()
    # get the list of corresponding labels
    labels = my_df['label'][idx_i].tolist()
    # and terms
    term = my_df['term'][idx_i].tolist()
    # and embeddings
    fastext_embeddings = my_df['fastext_embeddings'][idx_i].tolist()
    
    # compute the weight for each label between 0 and 1
    for j in range(len(idx_i)) : 
        l = labels[j]
        count_l = labels.count(l)
        #
        weight_l = count_l/len(idx_i)
        # print(i,term[j],"\n", l, weight_l)
        data_C.append([i, term[j],fastext_embeddings[j],l, weight_l])
        

my_df_C = pd.DataFrame(data_C)


my_df_C.columns = ['source', 'term', 'fastext_embeddings', 'label', 'label_weight']
my_df_C

import pickle
# save
my_df_C.to_pickle('./out/20220210_local_annot_accent_classif_FastTextEmb.pkl')

df_check = my_df_C.groupby('source')['term'].nunique()

len(df_check)
