==> 4.Get_BERT_Embeddings.ipynb <==
# GET BERT EMBEDDINGS
- Get Bert embeddings of extracted terms

import logging
#logging.basicConfig(level=logging.INFO)

import matplotlib.pyplot as plt


import numpy as np 
import pandas as pd 
import os
from sklearn.metrics import classification_report, confusion_matrix, multilabel_confusion_matrix, f1_score, fbeta_score, precision_score, recall_score, accuracy_score
import torch
import re

from transformers import CamembertModel, CamembertTokenizer, CamembertForSequenceClassification, AdamW, get_linear_schedule_with_warmup

os.environ['CUDA_VISIBLE_DEVICES'] = "1" 


import torch
from torch.utils.data import Dataset, TensorDataset, RandomSampler, SequentialSampler, DataLoader
from torch.nn import BCEWithLogitsLoss, Sigmoid


from sklearn.model_selection import train_test_split
from tqdm import tqdm


device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
print('device',device)

# the model : 
bert_name = "/export/home/cse200093/word-embedding/training-from-scratch-2021-08-13/"
# bert_name ="/export/home/opt/data/camembert/v0/camembert-base/"

lower = True
batch_size = 32
max_epochs = 50


Max_len = 128 
val_rate = .2
debug = 0

file_path = './out/15_02_22_CRH_VAL_FastText.pkl' 
OUT_FILE = './out/22_02_22_CRH_VAL_FastText_EDS_BERT.pkl' 

# Pre-processing of the text : removing the non-alphanumeric data : 
import unidecode

def strip(text, lower=False):
    # pattern = r"[^a-zA-z0-9\s,']"
    #text = unidecode.unidecode(text)
    text = re.sub(r'^-','',text, count = 1)
    if lower:
        return text.lower()
    else:
        return text


# load the df with fasttext embeddings
df = pd.read_pickle(file_path)
df['term'] = df['term'].apply(strip, args=(lower,)) 
df.sample(5)

# load the tokenizer : 
tokenizer = CamembertTokenizer.from_pretrained(bert_name, do_lower_case = lower)
train_texts = list(df['term'])


# tokenization of the train, val and test dataset : 
train_tokenizer_texts = tokenizer(train_texts, return_tensors = 'pt', padding=True, truncation=True, max_length = Max_len)
print('train_size ', train_tokenizer_texts['input_ids'].size())
attention_masks = train_tokenizer_texts['attention_mask']


indexed_tokens = tokenizer.convert_tokens_to_ids(tokenized_text)
segments_ids = [1] * len(tokenized_text)
tokens_tensor = torch.tensor([indexed_tokens])
segments_tensors = torch.tensor([segments_ids])

# Load pre-trained model (weights)
model = CamembertModel.from_pretrained(bert_name,
                                  output_hidden_states = True, # Whether the model returns all hidden-states.
                                  )

model.eval()
with torch.no_grad():
    outputs = model(tokens_tensor, segments_tensors)

print ("Number of layers:", len(hidden_states), "(initial embeddings + 12 BERT layers)")

layer_i = 0

print ("Number of batches:", len(hidden_states[layer_i]))
batch_i = 0
print ("Number of tokens:", len(hidden_states[layer_i][batch_i]))
token_i = 0
print ("Number of hidden units:", len(hidden_states[layer_i][batch_i][token_i]))

token_embeddings = torch.stack(hidden_states, dim=0)
token_embeddings.shape

# Remove dimension 1, the "batches".
token_embeddings = torch.squeeze(token_embeddings, dim=1)
token_embeddings.size()

# Swap dimensions 0 and 1.
token_embeddings = token_embeddings.permute(1,0,2)

token_embeddings.size()

# Stores the token vectors, with shape [22 x 3,072]
token_vecs_cat = []# `token_embeddings` is a [22 x 12 x 768] tensor.# For each token in the sentence...
for token in token_embeddings:
    
    # `token` is a [12 x 768] tensor# Concatenate the vectors (that is, append them together) from the last 
    # four layers.
    # Each layer vector is 768 values, so `cat_vec` is length 3,072.
    cat_vec = torch.cat((token[-1], token[-2], token[-3], token[-4]), dim=0)
    
    # Use `cat_vec` to represent `token`.
    token_vecs_cat.append(cat_vec)

print ('Shape is: %d x %d' % (len(token_vecs_cat), len(token_vecs_cat[0])))

# Stores the token vectors, with shape [22 x 768]
token_vecs_sum = []

# `token_embeddings` is a [22 x 12 x 768] tensor.

# For each token in the sentence...
for token in token_embeddings:

    # `token` is a [12 x 768] tensor

    # Sum the vectors from the last four layers.
    sum_vec = torch.sum(token[-4:], dim=0)
    
    # Use `sum_vec` to represent `token`.
    token_vecs_sum.append(sum_vec)

print ('Shape is: %d x %d' % (len(token_vecs_sum), len(token_vecs_sum[0])))


device = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
print('device', device)
batch_size = 16

BERT_embeddings = []

# Load pre-trained model (weights)
model = CamembertModel.from_pretrained(bert_name,
                                  output_hidden_states = True, # Whether the model returns all hidden-states.
                                  )

model.to(device)

dataset = TensorDataset(train_tokenizer_texts['input_ids'], train_tokenizer_texts['attention_mask'])
sampler = SequentialSampler(dataset)
dataloader = DataLoader(dataset, sampler=sampler, batch_size=batch_size)


model.eval()
for step, batch in enumerate(tqdm(dataloader)): 
    #Transfer batch to GPUs --> not adapted for EDS computation
    batch = tuple(t.to(device) for t in batch)
    #We dont need to update gradients as we are just predicting
    with torch.no_grad():
        #Bring up the next batch of input_texts and attention_masks 
        b_input_ids, b_input_mask = batch
        #Forward propogate the inputs and get output as logits
        outputs = model(b_input_ids, token_type_ids = None, attention_mask = b_input_mask)
        
        # Evaluating the model will return a different number of objects based on 
        # how it's  configured in the `from_pretrained` call earlier. In this case, 
        # becase we set `output_hidden_states = True`, the third item will be the 
        # hidden states from all layers. See the documentation for more details:
        # https://huggingface.co/transformers/model_doc/bert.html#bertmodel
        hidden_states = outputs[2]
        
        # To get a single vector for our entire sentence we have 
        # multiple application-dependent strategies, 
        # but a simple approach is to average the second to last 
        # hiden layer of each token producing a single 1024 length vector.
        # token_vecs = hidden_states[-2:][0]
        token_vecs = hidden_states[-2:][0]
        sentence_embedding = torch.mean(token_vecs, dim=1).to('cpu').numpy()
        BERT_embeddings.extend(sentence_embedding)
    


token_vecs_2 = hidden_states[-2][0]
print(token_vecs.shape, token_vecs_2.shape)
sentence_embedding_2 = torch.mean(token_vecs_2, dim=0)
print ("Our final sentence embedding vector of shape:", sentence_embedding_2.size())


len(BERT_embeddings), len(BERT_embeddings[0])

# Sanity checks : check if the embeddings of the same word 'SAPL' idx 8703 and 9214 have the same embeddings 
#BERT_embeddings[9214], BERT_embeddings[8703]

# check the similarity 
from scipy.spatial.distance import cosine
idx_1 = 6518
idx_2 = 9223
close_w = cosine(BERT_embeddings[idx_1], BERT_embeddings[idx_2])
print('similarity between',df['term'][idx_1],'and',df['term'][idx_2],'=', close_w)

# check the similarity 
from scipy.spatial.distance import cosine
idx_1 = 9223
idx_2 = 1468
close_w = cosine(BERT_embeddings[idx_1], BERT_embeddings[idx_2])
print('similarity between',df['term'][idx_1],'and',df['term'][idx_2],'=', close_w)

# check the similarity 
from scipy.spatial.distance import cosine
idx_1 = 4514
idx_2 = 12697
close_w = cosine(BERT_embeddings[idx_1], BERT_embeddings[idx_2])
print('similarity between',df['term'][idx_1],'and',df['term'][idx_2],'=', close_w)

# check the similarity 
from scipy.spatial.distance import cosine
idx_1 =10644
idx_2 = 13693
close_w = cosine(BERT_embeddings[idx_1], BERT_embeddings[idx_2])
print('similarity between "',df['term'][idx_1],'" and "',df['term'][idx_2],'" =', close_w)

# check the similarity 
from scipy.spatial.distance import cosine
idx_1 = 2043
idx_2 = 10644
close_w = cosine(BERT_embeddings[idx_1], BERT_embeddings[idx_2])
print('similarity between',df['term'][idx_1],'and',df['term'][idx_2],'=', close_w)

df.sample(30)

# add these BERT embeddings to our dataFrame 
# empty list of list to "append" each values
data = []
for i in range(len(df.values)):
    data.append([df.index[i], df.values[i,0],df.values[i,1],df.values[i,2],
                 df.values[i,3],
               BERT_embeddings[i],df.values[i,4]])
    
my_df = pd.DataFrame(data)
my_df.columns = ['idx','source', 'gender','term', 'fastext_embeddings', 'EDS_BERT_embeddings',
                 'label']

my_df[my_df.term == "ostéoporose"]

# save 
my_df.to_pickle(OUT_FILE)

my_df[my_df['term']=='ostéoporose'].EDS_BERT_embeddings

### TUTO BERT EMBEDDINGS
layer_i = 0
print ("Number of batches:", len(hidden_states[layer_i]))
batch_i = 0
print ("Number of tokens:", len(hidden_states[layer_i][batch_i]))
token_i = 0
print ("Number of hidden units:", len(hidden_states[layer_i][batch_i][token_i]))
# `hidden_states
print('      Type of hidden_states: ', type(hidden_states))
# Each layer in the list is a torch tensor.
print('Tensor shape for each layer: ', hidden_states[0].size())
# Concatenate the tensors for all layers. We use `stack` here to
# create a new dimension in the tensor.
token_embeddings = torch.stack(hidden_states, dim=0)
token_embeddings.size()
# `hidden_states` has shape [25x 1906 x 83 x 1024]
# To get a single vector for our entire sentence we have 
# multiple application-dependent strategies, 
# but a simple approach is to average the second to last 
# hiden layer of each token producing a single 768 length vector.

token_vecs = hidden_states[-2:][0]

# Calculate the average of all 22 token vectors.
sentence_embedding = torch.mean(token_vecs, dim=1)

token_vecs.size(), sentence_embedding.size()
for i, token_str in enumerate(train_tokenizer_texts):
    print (i, token_str)
