==> 6.Dist_matrix_computation.ipynb <==
# Distance matrix computation from an index patient 
A distance matrix of size (n_label, n_label) is computed between the index patient and all other patients. 

Distance computation for each label and cross label can be found in module distance.py

Input : 
A pickle file containing a dataframe of patients with following informations: 
source (file/clinical note id), patient_id, terms, Fastext Embedding, BERT embedding, label. 
Eventually : patient index file, if not, this one would be selected from the dataset


Output : Distance Matrix of size (n_label, n_label, number of patients to compare with)


import os
import numpy as np
import pandas as pd 

import sklearn
from scipy.spatial.distance import cosine
from sklearn.metrics import euclidean_distances
from sklearn.feature_extraction.text import CountVectorizer

# import local module of EMD computation along each label and between labels 
import distance 

file_path='./out/02022022_train_val_local_annotated_accent_lower_FastText_BERT_mean2l_Emb.pkl'
img_path = "./out/img_annot_Pheno_attributes/"
file_CLASS_path = './results/Gold_classif_expe_pheno_patient_id.csv'

Embedding = "Fasttext" # should be Fasttext of local_BERT or EDS_BERT
assert Embedding in ('Fasttext', 'local_BERT', 'EDS_BERT')

# load the dataframe
df_init = pd.read_pickle(file_path)

# in a first experiment, delete all "etatsosy rows"
df = df_init.copy()
df.drop(df.loc[df.label == 'etatsosy'].index, inplace=False)

# load the labels for all the terms of the file
labels = sorted(set(df['label']))
n_lab = len(labels)
print('\n number of labels', n_lab, '\nLabels', labels)



# load the files 
files = list(set(df['source']))
# print("number of files", len(files))

# Create a dictionnary  {file : [list of corresponding labels]} :
files_dict = {}
for i in range(len(files)):
    files_dict[files[i]]= set(df[df['source']== files[i]].label.tolist())
    
df['idx']=range(len(df))
df.sample(5)


# select a patient for comparison purpose : 
file_index = "CRH_scl_sample_46.ann"
files_lab = files.copy()
files_lab.remove(file_index)
print(len(files_lab))

# Compute Distance matrix 

# First by storing all distances in a dictionnary dist_dict : 
# dist_dict contains each following keys : (file, label_1, label_2) and value : distance 
dist_dict= {}
dist_matrix = np.zeros([len(files_lab), len(labels), len(labels)])


i = 0 
for file in sorted(set(files_lab)):
    j = 0
    for lab_1 in sorted(labels): 
        k = 0
        for lab_2 in sorted(labels): 
            if (file,lab_1,lab_2) not in dist_dict.keys(): 
                if lab_1 == lab_2:
                    dist_dict[(file, lab_1, lab_2)] = distance.distance_files_by_lab(df, file_index, file, 
                                                                                     Embedding,lab_1, 
                                                                                     Distance_type = 'cosine', 
                                                                                     verbose = False)
                    dist_matrix[i,j,j] = dist_dict[(file, lab_1, lab_2)]
                else : 
                    dist_dict[(file, lab_1, lab_2)] = distance.distance_files_cross_lab(df, file_index, file, 
                                                                                        Embedding,lab_1,lab_2,
                                                                                        Distance_type = 'cosine', 
                                                                                        verbose = False)
                    dist_matrix[i,j,k] = dist_dict[(file, lab_1, lab_2)]
            k += 1
        j += 1
                    
    i += 1
    if i%30 == 0:
        print('i', i)
    

# print(dist_dict[('CRH_lupus_sample_0.ann', 'etatsosy', 'osteomusculaires')])

# In order to easily navigate from one entity to another, create a dictionnary of labels and files 
# corresponding to their  
lab_index = {}
for i in range(len(labels)):
    lab_index[sorted(list(labels))[i]] = i

files_index ={}
for j in range(len(files_lab)):
    files_index[sorted(list(files_lab))[j]] = j
    
    
index_files={}
for j in range(len(files_lab)):
    index_files[j+1] = sorted(list(files_lab))[j]


# dist_matrix[files_index['CRH_lupus_sample_2.ann'], lab_index['etatsosy'], lab_index['immunitaire']]

# Now compute a weighted matrix in order to select only some labels of interest
# Get labels of interest : 
label_1 = 'respiratoire'
label_2 = 'immunitaire'


# get their respected indexes : 
idx_1 = lab_index[label_1]
idx_2 = lab_index[label_2]
# idx_3 = lab_index[label_3]

# give the associated weight : 
beta = 5

# build the weight matrix : 
weights = np.zeros([n_lab, n_lab]) # +100
weights[idx_1, idx_1] = weights[idx_2, idx_2] = 1
weights[idx_1, idx_2] = weights[idx_2, idx_1] = 0.5



# weights[0:idx_1+1, 0: idx_2+1]

# repeat the weights matrix for all patients : 
weights_all = np.repeat(weights[np.newaxis, :,:], len(files_lab), axis = 0)

weights_all.shape, weights_all[50, idx_1, idx_2]

# Now multiply the 2 3D matrices in order to get a distance from our patient to all patient on those 2 labels : 
print('Correponding shapes of the two matrices : distance matrix =', dist_matrix.shape, 
      'and weights =', weights_all.shape)

# we want to obtain a 150*1 array at the end

# Element wise computation 
result = np.multiply(dist_matrix, weights_all)

result.shape, np.nanmax(result)

result_no_nan = np.nan_to_num(result, copy=True, nan=0, posinf=None, neginf=None)

reduce_result = np.sum(result_no_nan, axis = 2)
reduce_result.shape

reduce_result_2 = np.sum(reduce_result, axis = 1)
reduce_result_2.shape

for file, index in files_index.items():
    if index == np.argmin(reduce_result_2):
        print(file)

# convert the np array to a dict in order to sort values and get the closest files
d = dict(enumerate(reduce_result_2.flatten(), 1))

# convert the number with file names :
file_dist_dict = {}
for k, v in d.items():
    file_dist_dict[index_files[k]] = v

#file_dist_dict

# rename the files to check if osteoporosis are the first one
# act_class = 'osteoporose'
# act_class = 'nephro_lupus'
act_class = 'PINS_Sclerodermie'

# Load the corresponding class of the files 
class_df = pd.read_csv(file_CLASS_path)
    
file_class_dict = {}

for k,v in file_dist_dict.items(): 
    if k in class_df['source'].tolist():
        if act_class in class_df[class_df['source'] == k].label.tolist():
            file_class_dict[(act_class[0:5]+'_'+k)] = v
        else :
            file_class_dict[k] = v

# class_df.sample(10)

file_index

{k: v for k, v in sorted(file_class_dict.items(), key=lambda item: item[1]) if v > 1}
